Remove superfluous dependencies, add plugin dependency
AspectJ only needs the runtime. AspectJ Maven however needs the tools,
if you want to ensure that the compiler uses the same version as the
runtime. Otherwise, the AspectJ version predefined for AspectJ Maven -
in this case 1.9.7 for the Mojohaus 1.14.0 version - is used for
compilation. The dev.aspectj version recommended by the AspectJ project
would use a more recent version by default, but not as recent as 1.9.19.

So, it is always advisable to override the plugin dependency. This will
become important later when upgrading to the AspectJ compiler snapshot
with the bugfix for this problem.